Add a deque of scene nodes and make them parents of arrow and covariance
Also, while here fix includes of rviz files.

Rationale behind this commit:

The arrow class like it's defined points to -Z, so we were rotating it to make
it point to +X. The problem is that if we make the arrow's node parent of
covariance's nodes we'll have the covariance rotated as well. The work-around is
to keep a deque of scene nodes, and make them parents of respective arrow and
covariance nodes. With this arrow's node can be rotated without affecting the
rotation of covariance's nodes.