[src] Refactor covariance visual nodes to have better control on orientation shapes offset
Now offset means position the 3D shapes on an axis or the height of the 2D
triangle for 2D poses.

The scale means an increase in the angle, but applied only in the shape and not
in the standard deviation. Thus only the scale of 1.0 represents one 1-sigma
standard deviation. Other scale values does not correspont to sigmas.